+++ tests/test_agent.py
@@ -371,10 +371,12 @@ def test_agent_update_agent_from_server(
     monkeypatch.setattr(
         server_fixture.__class__,
         "decrypt",
-        lambda self, encrypted_parameters: json.dumps([
-            {"name": "parameter1", "value": "new_value1", "is_environment": True},
-            {"name": "parameter2", "value": "new_value2", "is_environment": False},
-        ]),
+        lambda self, encrypted_parameters: json.dumps(
+            [
+                {"name": "parameter1", "value": "new_value1", "is_environment": True},
+                {"name": "parameter2", "value": "new_value2", "is_environment": False},
+            ]
+        ),
     )
     # Ensure supervisor_account is not None
     assert server_fixture.supervisor_account is not None
@@ -409,9 +411,9 @@ def test_agent_update_agent_from_server(
     monkeypatch.setattr(
         server_fixture.__class__,
         "decrypt",
-        lambda self, encrypted_parameters: json.dumps([
-            {"invalid_parameter": "invalid_value1"}
-        ]),
+        lambda self, encrypted_parameters: json.dumps(
+            [{"invalid_parameter": "invalid_value1"}]
+        ),
     )
     with pytest.raises(ValueError):
         agent_fixture.update_agent_from_server(server_fixture)